requirements.txt
# requirements.txt
tensorflow
numpy
scikit-learn
matplotlib
Pillow
imutils
streamlit
keras==2.15.0
opencv-python-headless==4.7.0.72